+++ changelog.md
@@ -3,6 +3,7 @@
 * **History**
  * Now tracks tier tokens received, group size and instanceMapIDs.
  * Tier tokens received is displayed in the more info details in the voting frame.
+ * Removed response text from voting frame more info to make it smaller.
 
 
 * Minor optimizations.